fix(ml): resolve type errors across ML modules
- anomaly.py: rename shadowed X/y variables in 3 loops to X_cv, X_sm, X_sup
- tasks.py: rename shadowed X/y to X_t, y_t
- ensemble.py: fix undefined max_windows → self.max_windows
- federated.py: fix None attr access with assert, fix union-attr on extend
- cross_stream.py: add Any type annotation to ATTACK_SEQUENCES dict
- dataset_adapters/base.py: cast heuristic_label result to int for TypedDict
- dataset_import.py: add type ignore for abstract BaseAdapter instantiation
- deep_features.py: rename shadowed intervals variable
- comparison.py: annotate best_score as float
- community_rules.py: add type annotations to undeclared variables
- ueba.py: fix hosts.get() call, add annotations
- 18 files, ~30 errors fixed

+++ backend/ml/dataset_adapters/base.py
@@ -127,7 +127,7 @@ def shannon_entropy(text: str) -> float:
     return min(1.0, entropy / 7.0)
 
 
-def heuristic_label(event: NormalizedEventDict) -> int:
+def heuristic_label(event: NormalizedEventDict) -> float:
     """Apply BARAQ's heuristic labeling to an event.
 
     Returns 1 (attack) or 0 (benign).  Analyses use this for unsupervised
@@ -235,7 +235,7 @@ class BaseAdapter(ABC):
                 seen.add(fp)
                 # Apply heuristic label if not already labeled
                 if "label" not in parsed:
-                    parsed["label"] = heuristic_label(parsed)
+                    parsed["label"] = int(heuristic_label(parsed))
                 events.append(parsed)
             except Exception as exc:
                 errors.append(f"event {total}: {exc}")
